fix lint, moved step
