+++ features/purpose/purpose-listing-consumer.feature
@@ -7,7 +7,7 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "<ruolo>" di "<ente>"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "<ente>" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "<ente>" ha già creato 5 finalità in stato "ACTIVE" per quell'eservice
+    Given "<ente>" ha già creato 5 finalità in stato "ACTIVE" per quell'eservice
     When l'utente fruitore richiede una operazione di listing delle finalità limitata ai primi 3 risultati
     Then si ottiene status code <risultato> e la lista di 3 finalità
 
@@ -34,7 +34,7 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "admin" di "PA1"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "PA1" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "PA1" ha già creato 5 finalità in stato "ACTIVE" per quell'eservice
+    Given "PA1" ha già creato 5 finalità in stato "ACTIVE" per quell'eservice
     When l'utente fruitore richiede una operazione di listing delle finalità con offset 2
     Then si ottiene status code 200 e la lista di 3 finalità
 
@@ -43,7 +43,7 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "admin" di "PA1"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "PA1" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "PA1" ha già creato 2 finalità in stato "ACTIVE" per quell'eservice
+    Given "PA1" ha già creato 2 finalità in stato "ACTIVE" per quell'eservice
     When l'utente fruitore richiede una operazione di listing delle sue finalità sugli e-services a cui è sottoscritto
     Then si ottiene status code 200 e la lista di 2 finalità
 
@@ -52,10 +52,10 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "admin" di "GSP"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "PA1" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "PA1" ha già creato 1 finalità in stato "ACTIVE" per quell'eservice
+    Given "PA1" ha già creato 1 finalità in stato "ACTIVE" per quell'eservice
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "GSP" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "GSP" ha già creato 1 finalità in stato "ACTIVE" per quell'eservice
+    Given "GSP" ha già creato 1 finalità in stato "ACTIVE" per quell'eservice
     When l'utente fruitore richiede una operazione di listing delle finalità filtrata per il secondo e-service
     Then si ottiene status code 200 e la lista di 1 finalità
 
@@ -64,8 +64,8 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "admin" di "PA1"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "PA1" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "PA1" ha già creato 1 finalità in stato "ACTIVE" per quell'eservice
-    Given un "admin" di "PA1" ha già creato 1 finalità in stato "SUSPENDED" per quell'eservice
+    Given "PA1" ha già creato 1 finalità in stato "ACTIVE" per quell'eservice
+    Given "PA1" ha già creato 1 finalità in stato "SUSPENDED" per quell'eservice
     When l'utente fruitore richiede una operazione di listing delle finalità in stato "<statoFinalita>"
     Then si ottiene status code 200 e la lista di 1 finalità
 
@@ -79,7 +79,7 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "admin" di "PA1"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "PA1" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "PA1" ha già creato una finalità in stato "ACTIVE" per quell'e-service contenente la keyword "test"
+    Given "PA1" ha già creato una finalità in stato "ACTIVE" per quell'e-service contenente la keyword "test"
     When l'utente fruitore richiede una operazione di listing delle finalità filtrando per la keyword "test"
     Then si ottiene status code 200 e la lista di 1 finalità
 
@@ -88,6 +88,6 @@ Feature: Lista delle finalità lato fruitore
     Given l'utente è un "admin" di "PA1"
     Given un "admin" di "PA2" ha già creato e pubblicato 1 e-service
     Given "PA1" ha una richiesta di fruizione in stato "ACTIVE" per quell'e-service
-    Given un "admin" di "PA1" ha già creato una finalità in stato "ACTIVE" per quell'e-service contenente la keyword "test"
+    Given "PA1" ha già creato una finalità in stato "ACTIVE" per quell'e-service contenente la keyword "test"
     When l'utente fruitore richiede una operazione di listing delle finalità filtrando per la keyword "unknown"
     Then si ottiene status code 200 e la lista di 0 finalità
